+++ chunk.h
@@ -26,15 +26,15 @@ enum OpCode : uint8_t
 class Chunk
 {
 public:
-    void write(uint8_t byte, int line);
+    void write(uint8_t byte, size_t line);
     size_t addConstant(Value constant);
     const std::vector<uint8_t>& getCode() const { return code; }
-    const std::vector<int>& getLines() const { return lines; }
+    const std::vector<size_t>& getLines() const { return lines; }
     const std::vector<Value>& getConstants() const { return constants; }
 
 private:
     std::vector<uint8_t> code;
-    std::vector<int> lines;
+    std::vector<size_t> lines;
     std::vector<Value> constants;
 };
